feat: DB-backed settings with hot-reload
Move HTTP client config (timeout, retries, user agent) and feed poll
interval out of `wyrm.toml` and into a singleton `settings` table. Settings
are loaded at startup into an `Arc<RwLock<RuntimeSettings>>` shared between
the actix handlers and the RSS worker threads.

- Add settings migration and Diesel model with singleton constraint
- Refactor HttpClient
- Worker re-reads `feed_poll_interval_secs` each loop iteration and accepts
  a `Reconfigure` command to rebuild the HTTP client without restarting
- Settings API route + CRUD handler: persists to DB, updates runtime_settings
  in place, then sends Reconfigure to interrupt the worker's current sleep
- Rename `utils/src/settings.rs` -> `config.rs` to make clear it only handles
  startup/file-based config (`bind`, `port`, database URL)
- Frontend: Settings/UpdateSettings types, useSettings hook, General
  settings component wired into the settings page

+++ database/src/views/post.rs
@@ -30,7 +30,7 @@ impl PostQuery {
     pub async fn list(
         self,
         pool: &DatabasePool,
-        page_size: i64,
+        page_size: i32,
     ) -> WyrmResult<PagedResponse<Vec<Post>>> {
         let mut conn = pool.get().await?;
 
@@ -39,7 +39,7 @@ impl PostQuery {
         let mut query = posts::table
             .select(Post::as_select())
             .order((posts::published_at.desc(), posts::id.desc()))
-            .limit(page_size + 1)
+            .limit(page_size as i64 + 1)
             .into_boxed();
 
         if let Some(feed_id) = self.feed_id {
@@ -69,7 +69,7 @@ impl PostQuery {
         }
 
         let mut items: Vec<Post> = query.load(&mut conn).await.map_err(WyrmError::from)?;
-        let next_page = if items.len() as i64 > page_size {
+        let next_page = if items.len() > page_size as usize {
             items.truncate(page_size as usize);
             items
                 .last()
